[P24-17] feat: 사용자 관련 Controller, Service 클래스명 변경
