+++ BE/src/main/java/com/example/p24zip/domain/user/controller/AuthController.java
@@ -1,4 +1,30 @@
 package com.example.p24zip.domain.user.controller;
 
+import com.example.p24zip.domain.user.dto.request.SignupRequestDto;
+import com.example.p24zip.domain.user.service.AuthService;
+import com.example.p24zip.global.response.ApiResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/auth")
 public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<ApiResponse<Void>> signup(
+        @RequestBody @Valid SignupRequestDto requestDto) {
+        authService.signup(requestDto);
+
+        return ResponseEntity.ok(
+            ApiResponse.ok("CREATED", "회원가입을 성공했습니다.", null)
+        );
+    }
 }
