--- BE/src/main/java/com/example/p24zip/domain/user/controller/UserController.java
@@ -1,30 +1,0 @@
-package com.example.p24zip.domain.user.controller;
-
-import com.example.p24zip.domain.user.dto.request.SignupRequestDto;
-import com.example.p24zip.domain.user.service.UserService;
-import com.example.p24zip.global.response.ApiResponse;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/auth")
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<ApiResponse<Void>> signup(
-        @RequestBody @Valid SignupRequestDto requestDto) {
-        userService.signup(requestDto);
-
-        return ResponseEntity.ok(
-            ApiResponse.ok("CREATED", "회원가입을 성공했습니다.", null)
-        );
-    }
-}
